backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
pydantic
fastapi-cors
python-multipart
Pillow 